onprem-ml-baseline/requirements.txt
# Web Framework
Flask
flask-cors
Werkzeug

# Database
psycopg2-binary

# Environment Variables
python-dotenv

# System Monitoring
psutil

# HTTP Client (for testing)
requests

# Dependencies
click
blinker
itsdangerous
Jinja2
MarkupSafe

